amalgamation
git-svn-id: file:///home/notaz/opt/svn/PicoDrive@227 be3aeb3a-fb24-0410-a615-afba39da0efa

+++ Pico/Patch.c
@@ -213,7 +213,7 @@ bad_code:
 
 
 
-unsigned short PicoRead16(unsigned int a);
+unsigned int PicoRead16(unsigned int a);
 void PicoWrite16(unsigned int a, unsigned short d);
